rename functional-test-types to e2e-types
