+++ extensions/amp-sidebar/1.0/test-e2e/test-amp-sidebar.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 //
-import {Key} from '../../../../build-system/tasks/e2e/functional-test-types';
+import {Key} from '../../../../build-system/tasks/e2e/e2e-types';
 
 describes.endtoend(
   'amp-sidebar',
